pyproject.toml
[project]
name = "jira-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "atlassian-python-api>=3.41.0",
    "fastapi>=0.115.0",
    "fastmcp>=3.2.0",
    "jira>=3.10.5",
    "mcp>=1.27.0",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.34.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
]
